Refactoring: - better error handling - unnesessary loops removed - state storage added

+++ misc/sql/storage/migration.sql
@@ -3,7 +3,6 @@ create table files
     name varchar not null
         primary key
         unique,
-    size int,
+    size bigint,
     data bytea   not null
 );
-
